+++ lib/contentful/management/space_api_key_methods_factory.rb
@@ -1,0 +1,13 @@
+require_relative 'space_association_methods_factory'
+
+module Contentful
+  module Management
+    class SpaceApiKeyMethodsFactory
+      include Contentful::Management::SpaceAssociationMethodsFactory
+
+      def new
+        fail 'Not supported'
+      end
+    end
+  end
+end